requirements.txt
asgiref==3.6.0
boto3==1.26.135
botocore==1.29.135
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.2
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
